Renaming, some minor rewrites, and re-added images from the report
Eventually a git fat or git large file service or so on link can return, but for the moment these files will return as bmp to make sure they're accessible for that transition later on.
