+++ 2d/constantSlopeBeach/Ting_and_Kirby_breaking/Ting_and_Kirby_breaking_so.py
@@ -1,7 +1,7 @@
-import tank
 from proteus.default_so import *
+import Ting_and_Kirby_breaking
 
-if tank.useOnlyVF:
+if Ting_and_Kirby_breaking.useOnlyVF:
     pnList = [("twp_navier_stokes_p", "twp_navier_stokes_n"),
               ("vof_p",               "vof_n")]
 else:
@@ -12,14 +12,14 @@ else:
               ("ls_consrv_p",         "ls_consrv_n")]
     
     
-if tank.useRANS > 0:
+if Ting_and_Kirby_breaking.useRANS > 0:
     pnList.append(("kappa_p",
                    "kappa_n"))
     pnList.append(("dissipation_p",
                    "dissipation_n"))
 name = "tank_p" 
 
-if tank.timeDiscretization == 'flcbdf':
+if Ting_and_Kirby_breaking.timeDiscretization == 'flcbdf':
     systemStepControllerType = Sequential_MinFLCBDFModelStep
     systemStepControllerType = Sequential_MinAdaptiveModelStep
 else:
@@ -28,5 +28,5 @@ else:
 needEBQ_GLOBAL = False
 needEBQ = False
 
-tnList = [0.0, tank.dt_init] + [i * tank.dt_fixed for i in range(1,
-                                                                 tank.nDTout + 1)]
+tnList = [0.0, Ting_and_Kirby_breaking.dt_init] + [i * Ting_and_Kirby_breaking.dt_fixed for i in range(1, Ting_and_Kirby_breaking.nDTout + 1)]
+archiveFlag = ArchiveFlags.EVERY_SEQUENCE_STEP
