+++ 2d/constantSlopeBeach/Ting_and_Kirby_breaking/ls_consrv_n.py
@@ -1,5 +1,5 @@
 from proteus import *
-from tank import *
+from Ting_and_Kirby_breaking import *
 from ls_consrv_p import *
 
 timeIntegrator  = ForwardIntegrator
