+++ 2d/constantSlopeBeach/Ting_and_Kirby_breaking/twp_navier_stokes_n.py
@@ -1,6 +1,6 @@
 from proteus import *
 from twp_navier_stokes_p import *
-from tank import *
+from Ting_and_Kirby_breaking import *
 
 if timeDiscretization=='vbdf':
     timeIntegration = VBDF
