+++ 2d/constantSlopeBeach/Xie_vertical_breakwater/ls_consrv_n.py
@@ -1,5 +1,5 @@
 from proteus import *
-from tank import *
+from Xie_vertical_breakwater import *
 from ls_consrv_p import *
 
 timeIntegrator  = ForwardIntegrator
