+++ 2d/constantSlopeBeach/Xie_vertical_breakwater/redist_p.py
@@ -1,7 +1,7 @@
 from proteus import *
 from proteus.default_p import *
 from math import *
-from tank import *
+from Xie_vertical_breakwater import *
 from proteus.mprans import RDLS
 """
 The redistancing equation in the sloshbox test problem.
