+++ 2d/constantSlopeBeach/linear_wave_shoaling/linear_wave_shoaling_so.py
@@ -1,7 +1,7 @@
 from proteus.default_so import *
-import tank
+import linear_wave_shoaling
 
-if tank.useOnlyVF:
+if linear_wave_shoaling.useOnlyVF:
     pnList = [("twp_navier_stokes_p", "twp_navier_stokes_n"),
               ("vof_p",               "vof_n")]
 else:
@@ -12,14 +12,14 @@ else:
               ("ls_consrv_p",         "ls_consrv_n")]
     
     
-if tank.useRANS > 0:
+if linear_wave_shoaling.useRANS > 0:
     pnList.append(("kappa_p",
                    "kappa_n"))
     pnList.append(("dissipation_p",
                    "dissipation_n"))
 name = "tank_p" 
 
-if tank.timeDiscretization == 'flcbdf':
+if linear_wave_shoaling.timeDiscretization == 'flcbdf':
     systemStepControllerType = Sequential_MinFLCBDFModelStep
     systemStepControllerType = Sequential_MinAdaptiveModelStep
 else:
@@ -28,5 +28,5 @@ else:
 needEBQ_GLOBAL = False
 needEBQ = False
 
-tnList = [0.0,tank.dt_init]+[i*tank.dt_fixed for i in range(1,tank.nDTout+1)] 
+tnList = [0.0, linear_wave_shoaling.dt_init] + [i * linear_wave_shoaling.dt_fixed for i in range(1, linear_wave_shoaling.nDTout + 1)]
 archiveFlag = ArchiveFlags.EVERY_SEQUENCE_STEP
