+++ 2d/constantSlopeBeach/linear_wave_shoaling/ls_p.py
@@ -1,6 +1,6 @@
 from proteus import *
 from proteus.default_p import *
-from tank import *
+from linear_wave_shoaling import *
 from proteus.mprans import NCLS
 
 LevelModelType = NCLS.LevelModel
