+++ 2d/constantSlopeBeach/linear_wave_shoaling/redist_p.py
@@ -1,7 +1,7 @@
 from proteus import *
 from proteus.default_p import *
 from math import *
-from tank import *
+from linear_wave_shoaling import *
 from proteus.mprans import RDLS
 """
 The redistancing equation in the sloshbox test problem.
