+++ 2d/constantSlopeBeach/linear_wave_shoaling/vof_p.py
@@ -1,7 +1,7 @@
 from proteus import *
 from proteus.default_p import *
 from proteus.ctransportCoefficients import smoothedHeaviside
-from tank import *
+from linear_wave_shoaling import *
 from proteus.mprans import VOF
 
 LevelModelType = VOF.LevelModel
